First Workings
Plotted Anscombe #1 using code from in-class tutorial with as few changes as possible. Found out how to reference a DIV id instead of appending to BODY. All other changes were straightforward. Changed X axis specification to be 0-MAX instead of EXTENT becuase it was causing a problem with a reversed range going 10-9 only.